ai_service/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0